Fix tests

+++ pandas/tests/test_algos.py
@@ -248,13 +248,13 @@ class TestFactorize:
 
     def test_factorizer_with_mask(self):
         # GH#49549
-        data = np.array([1, 2, 3, 1, 1, 0])
+        data = np.array([1, 2, 3, 1, 1, 0], dtype="int64")
         mask = np.array([False, False, False, False, False, True])
         rizer = ht.Int64Factorizer(len(data))
         result = rizer.factorize(data, mask=mask)
-        expected = np.array([0, 1, 2, 0, 0, -1])
+        expected = np.array([0, 1, 2, 0, 0, -1], dtype="int64")
         tm.assert_numpy_array_equal(result, expected)
-        expected_uniques = np.array([1, 2, 3])
+        expected_uniques = np.array([1, 2, 3], dtype="int64")
         tm.assert_numpy_array_equal(rizer.uniques.to_array(), expected_uniques)
 
     def test_factorizer_object_with_nan(self):
@@ -262,7 +262,7 @@ class TestFactorize:
         data = np.array([1, 2, 3, 1, np.nan])
         rizer = ht.ObjectFactorizer(len(data))
         result = rizer.factorize(data.astype(object))
-        expected = np.array([0, 1, 2, 0, -1])
+        expected = np.array([0, 1, 2, 0, -1], dtype="int64")
         tm.assert_numpy_array_equal(result, expected)
         expected_uniques = np.array([1, 2, 3], dtype=object)
         tm.assert_numpy_array_equal(rizer.uniques.to_array(), expected_uniques)
